+++ terra_geocrud/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'terra_geocrud.apps.TerraCrudConfig'