+++ terra_geocrud/tests/test_admin.py
@@ -2,8 +2,8 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
 from rest_framework import status
-from geostore.models import Layer
 
+from geostore.models import Layer
 from .. import models
 
 User = get_user_model()